switch from github workflows to github actions workflows
