+++ services/github/github-workflow-status.tester.js
@@ -1,43 +1,42 @@
-import Joi from 'joi'
-import { isBuildStatus } from '../build-status.js'
-import { createServiceTester } from '../tester.js'
-export const t = await createServiceTester()
+import { ServiceTester } from '../tester.js'
 
-const isWorkflowStatus = Joi.alternatives()
-  .try(isBuildStatus, Joi.equal('no status'))
-  .required()
+export const t = new ServiceTester({
+  id: 'GithubWorkflowStatus',
+  title: 'Github Workflow Status',
+  pathPrefix: '/github/workflow/status',
+})
 
-t.create('nonexistent repo')
+t.create('no longer available (previously nonexistent repo)')
   .get('/badges/shields-fakeness/fake.json')
   .expectBadge({
     label: 'build',
-    message: 'repo, branch, or workflow not found',
+    message: 'no longer available',
   })
 
-t.create('nonexistent workflow')
+t.create('no longer available (previously nonexistent workflow)')
   .get('/actions/toolkit/not-a-real-workflow.json')
   .expectBadge({
     label: 'build',
-    message: 'repo, branch, or workflow not found',
+    message: 'no longer available',
   })
 
-t.create('valid workflow')
+t.create('no longer available (previously valid workflow)')
   .get('/actions/toolkit/toolkit-unit-tests.json')
   .expectBadge({
     label: 'build',
-    message: isWorkflowStatus,
+    message: 'no longer available',
   })
 
-t.create('valid workflow (branch)')
+t.create('no longer available (previously valid workflow - branch)')
   .get('/actions/toolkit/toolkit-unit-tests/master.json')
   .expectBadge({
     label: 'build',
-    message: isWorkflowStatus,
+    message: 'no longer available',
   })
 
-t.create('valid workflow (event)')
+t.create('no longer available (previously valid workflow - event)')
   .get('/actions/toolkit/toolkit-unit-tests.json?event=push')
   .expectBadge({
     label: 'build',
-    message: isWorkflowStatus,
+    message: 'no longer available',
   })
